+++ lib/src/analyzers/lint_analyzer/rules/rules_factory.dart
@@ -35,6 +35,7 @@ import 'rules_list/double_literal_format/double_literal_format_rule.dart';
 import 'rules_list/format_comment/format_comment_rule.dart';
 import 'rules_list/member_ordering/member_ordering_rule.dart';
 import 'rules_list/member_ordering_extended/member_ordering_extended_rule.dart';
+import 'rules_list/missing_test_assertion/missing_test_assertion_rule.dart';
 import 'rules_list/newline_before_return/newline_before_return_rule.dart';
 import 'rules_list/no_boolean_literal_compare/no_boolean_literal_compare_rule.dart';
 import 'rules_list/no_empty_block/no_empty_block_rule.dart';
@@ -109,6 +110,7 @@ final _implementedRules = <String, Rule Function(Map<String, Object>)>{
   FormatCommentRule.ruleId: FormatCommentRule.new,
   MemberOrderingRule.ruleId: MemberOrderingRule.new,
   MemberOrderingExtendedRule.ruleId: MemberOrderingExtendedRule.new,
+  MissingTestAssertionRule.ruleId: MissingTestAssertionRule.new,
   NewlineBeforeReturnRule.ruleId: NewlineBeforeReturnRule.new,
   NoBooleanLiteralCompareRule.ruleId: NoBooleanLiteralCompareRule.new,
   NoEmptyBlockRule.ruleId: NoEmptyBlockRule.new,