+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 4.21.0
+## Unreleased
 
 * feat: add static code diagnostic [`missing-test-assertion`](https://dartcodemetrics.dev/docs/rules/common/missing-test-assertion).
+* feat: add support for presets
 
 ## 4.20.0